Fixed Puffish attribute skills not working for 2 reasons #1 - they used the "puffish_skills" modid, when the attributes are from "puffish_attributes" #2 - the mod is weird and has no base value for them, while most of the skills were using "multiply" operations. 0 * 0.12 = 0, as i'm sure you know.
renamed roll recharge skill internally and externally, also fixed/buffed it to use "multiply_base", because the attribute has a base value of 20, which was getting 0.10 added to it, not very good.
renamed roll charges skill to not have the floating point because it isn't relevant, makes it more confusing with the recharge arguably.